summit rcm: preserve permanent on factory reset
Mutual cert-based authentication requires factory-
provisioned files and timestamp data. Prevent these
files from being removed on factory resets

PROD-26075

+++ package/summit-encrypted-storage-toolkit/rootfs/usr/sbin/do_factory_reset.sh
@@ -25,7 +25,13 @@ do_check_and_reset() {
 	if [ -f "${RESET_INIDICATOR}" ]; then
 		# Delete all user data, but not the /data/secret dir as it is encrypted.
 		find /data -maxdepth 1 -mindepth 1 ! -name secret -exec rm -fr {} \;
-		find ${USER_SETTINGS_SECRET_TARGET} -maxdepth 1 -mindepth 1 ! -name permanent -exec rm -fr {} \;
+		if grep -q 'enable_client_auth:\s*True' /etc/summit-rcm.ini 2>/dev/null; then
+			# Preserve factory provisioned files and timestamp data
+			find "${USER_SETTINGS_SECRET_TARGET:?}" -maxdepth 1 -mindepth 1 ! -name permanent -exec rm -fr {} \;
+		else
+			rm -fr "${USER_SETTINGS_SECRET_TARGET:?}/"*
+		fi
+
 		# Run factory reset hooks for external components
 		for hook_sh in /usr/sbin/factory_reset_*.sh; do
 			# shellcheck source=/dev/null
